PEAR-2035 move dictionary viewer to new location
Previously, the dictionary viewer was implememted as part of a custom theme. This does not appear
to be how mkdocs is intended to work. This PR moved the dict viewer into the docs directory,
which may not be completely idomatic either.

As of this commit, the dict viewer works again, but it's styling needs to be updated still.