PEAR-2035 re-adding the encyclopedia index
The buildencyclopedia.py script removed the Index entry from
the encyclopedia entries in the mkdocs.yml config. This removes
the browsable encyclopedia index.  This comment re-adds the index
and comments out the buildencyclopedia script from the deploy script.

We should probably change the process and make the build encyclopedia
script part of the development process instead of the build and deploy
process.