PEAR-2352 remove mvs app and update links
The gdcmvs app can be removed from the doc site. The MVS hosted
tool will be used instead. The remaining links are updated. A
similar html injection technique was needed for the of the links.
In markdown files, we were able to use the attr_list extension.
This would be the preferred approach.