Pear 2352 mvs search (#469)
* PEAR-2352 remove mvs app and update links

In order for the link to open in a new tab, I had to use HTML
injection. This happens at compile-time, not at run-time. It is
possible that this breaks if mkdocs changing their site generation
to escape HTML.

The gdcmvs app can be removed from the doc site. The MVS hosted
tool will be used instead. The remaining links are updated. A
similar html injection technique was needed for the of the links.
In markdown files, we were able to use the attr_list extension.
This would be the preferred approach.